Convert FileWidget into a class
Result of fileWidgets are handled similarly. So, this is to prevent code duplication
Later File Chooser intents can also be added to the same class
