--- collect_app/src/main/java/org/odk/collect/android/widgets/interfaces/FileWidget.java
@@ -1,8 +1,0 @@
-package org.odk.collect.android.widgets.interfaces;
-
-/**
- * @author James Knight
- */
-public interface FileWidget extends BinaryWidget {
-    void deleteFile();
-}
